Monte carlo simulation
Modified the bayes script so that if you are executing it directly, it functions as the original author intended. However, if you run the accompanying script, it will execute a monte carlo simulation of whatever decision algorithm you pass it. It can then tell you how efficient on average the algorithm is. Currently hardcoded to do so for choosing the highest or splitting between the two highest probability search areas.